LayoutTests/fast/css/calc-parsing.html

<!DOCTYPE html>
<html>
    <body>
        <script src="../../resources/js-test-pre.js"></script>
        <div id="testDiv"></div>
        <script>
            description("Tests parsing of various valid and invalid calc expressions.");

            var testDiv = document.getElementById("testDiv");
            
            function testExpression(expression, specifiedValue, computedValue)
            {
                debug('');
                testDiv.style["width"] = '999px';

                evalAndLog(`testDiv.style["width"] = "${expression}"`);
                shouldBeEqualToString("testDiv.style['width']", `${specifiedValue}`);
                shouldBeEqualToString("window.getComputedStyle(testDiv).getPropertyValue('width')", `${computedValue}`);
            }

            // Valid expressions.
            testExpression('calc(100px)', 'calc(100px)', '100px');
            testExpression('max(100px + 200px)', 'max(300px)', '300px');
            testExpression('max(100px , 200px)', 'max(200px)', '200px');
            testExpression('max(100px,200px)', 'max(200px)', '200px');

            // Non-parsing expression.
            testExpression('calc(100px, 200px)', '999px', '999px');
            testExpression('calc(100px  200px)', '999px', '999px');

            testExpression('calc(100px ( 200px)', '999px', '999px');

            testExpression('min(100px 200px)', '999px', '999px');
            testExpression('max(100px 200px)', '999px', '999px');
            testExpression('max(100px,, 200px)', '999px', '999px');
            testExpression('max(100px, , 200px)', '999px', '999px');
            testExpression('max(100px, 200px,)', '999px', '999px');
        </script>
        <script src="../../resources/js-test-post.js"></script>
    </body>
</html>
