LayoutTests/fast/css/calc-parsing-expected.txt

Tests parsing of various valid and invalid calc expressions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



testDiv.style["width"] = "calc(100px)"
PASS testDiv.style['width'] is "calc(100px)"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "100px"

testDiv.style["width"] = "max(100px + 200px)"
PASS testDiv.style['width'] is "max(300px)"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "300px"

testDiv.style["width"] = "max(100px , 200px)"
PASS testDiv.style['width'] is "max(200px)"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "200px"

testDiv.style["width"] = "max(100px,200px)"
PASS testDiv.style['width'] is "max(200px)"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "200px"

testDiv.style["width"] = "calc(100px, 200px)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"

testDiv.style["width"] = "calc(100px  200px)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"

testDiv.style["width"] = "calc(100px ( 200px)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"

testDiv.style["width"] = "min(100px 200px)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"

testDiv.style["width"] = "max(100px 200px)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"

testDiv.style["width"] = "max(100px,, 200px)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"

testDiv.style["width"] = "max(100px, , 200px)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"

testDiv.style["width"] = "max(100px, 200px,)"
PASS testDiv.style['width'] is "999px"
PASS window.getComputedStyle(testDiv).getPropertyValue('width') is "999px"
PASS successfullyParsed is true

TEST COMPLETE
